Created separate maven modules
