+++ dragonball-cli/src/main/kotlin/me/vendoor/dragonball/cli/commands/Migrate.kt
@@ -1,0 +1,33 @@
+package me.vendoor.dragonball.cli.commands
+
+import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.parameters.options.option
+import com.github.ajalt.clikt.parameters.options.required
+import com.github.ajalt.clikt.parameters.types.file
+import me.vendoor.dragonball.library.migration.MigrationPerformer
+import me.vendoor.dragonball.common.database.openClient
+import me.vendoor.dragonball.cli.util.loadConfigurationFrom
+import java.io.File
+
+class Migrate: CliktCommand(
+        help = "Migrates an initialized database."
+) {
+    private val configFile: File by option(
+            help = "Path to the configuration file."
+    ).file().required()
+
+    private val targetVersion: String by option(
+            help = "The target version of the migration"
+    ).required()
+
+    override fun run() {
+        val configuration = loadConfigurationFrom(configFile)
+
+        val client = openClient(configuration.database.connectionString)
+        val database = client.getDatabase(configuration.database.name)
+
+        MigrationPerformer(database).perform(targetVersion)
+
+        client.close()
+    }
+}
