+++ dragonball-common/src/main/kotlin/me/vendoor/dragonball/common/database/IndexSort.kt
@@ -1,0 +1,6 @@
+package me.vendoor.dragonball.common.database
+
+enum class IndexSort(val intValue: Int) {
+    ASCENDING(1),
+    DESCENDING(-1)
+}
