+++ dragonball-library/src/main/kotlin/me/vendoor/dragonball/library/configuration/Configuration.kt
@@ -1,0 +1,11 @@
+package me.vendoor.dragonball.library.configuration
+
+class Configuration {
+    // Abusing lateinit? Maybe.
+    lateinit var database: Database
+
+    class Database {
+        lateinit var connectionString: String
+        lateinit var name: String
+    }
+}
