+++ dragonball-library/src/main/kotlin/me/vendoor/dragonball/library/migration/MigrationScript.kt
@@ -1,0 +1,8 @@
+package me.vendoor.dragonball.library.migration
+
+abstract class MigrationScript(
+        val version: String,
+        val description: String
+) {
+    abstract fun migrate(context: MigrationContext)
+}
