+++ dragonball-library/src/main/kotlin/me/vendoor/dragonball/library/util/time/TimeSource.kt
@@ -1,0 +1,9 @@
+package me.vendoor.dragonball.library.util.time
+
+import java.time.Clock
+
+object TimeSource {
+    private val clock = Clock.systemUTC()
+
+    fun currentTimestamp() = clock.millis()
+}
