--- src/main/kotlin/me/vendoor/dragonball/api/configuration/Configuration.kt
@@ -1,11 +1,0 @@
-package me.vendoor.dragonball.api.configuration
-
-class Configuration {
-    // Abusing lateinit? Maybe.
-    lateinit var database: Database
-
-    class Database {
-        lateinit var connectionString: String
-        lateinit var name: String
-    }
-}
