--- src/main/kotlin/me/vendoor/dragonball/api/migration/MigrationScript.kt
@@ -1,8 +1,0 @@
-package me.vendoor.dragonball.api.migration
-
-abstract class MigrationScript(
-        val version: String,
-        val description: String
-) {
-    abstract fun migrate(context: MigrationContext)
-}
