--- src/main/kotlin/me/vendoor/dragonball/api/migration/MigrationScriptRegistry.kt
@@ -1,12 +1,0 @@
-package me.vendoor.dragonball.api.migration
-
-object MigrationScriptRegistry {
-    private val scripts: MutableList<MigrationScript> = ArrayList()
-
-    fun getRegisteredScripts(): List<MigrationScript> =
-            scripts
-
-    fun registerMigrationScript(script: MigrationScript) {
-        scripts.add(script)
-    }
-}
