--- src/main/kotlin/me/vendoor/dragonball/api/util/database/IndexSort.kt
@@ -1,6 +1,0 @@
-package me.vendoor.dragonball.api.util.database
-
-enum class IndexSort(val intValue: Int) {
-    ASCENDING(1),
-    DESCENDING(-1)
-}
