--- src/main/kotlin/me/vendoor/dragonball/api/util/time/TimeSource.kt
@@ -1,9 +1,0 @@
-package me.vendoor.dragonball.api.util.time
-
-import java.time.Clock
-
-object TimeSource {
-    private val clock = Clock.systemUTC()
-
-    fun currentTimestamp() = clock.millis()
-}
