--- src/main/kotlin/me/vendoor/dragonball/specification/VendoorDatabaseSpecification.kt
@@ -1,13 +1,0 @@
-package me.vendoor.dragonball.specification
-
-import me.vendoor.dragonball.api.dsl.upsert.database
-
-fun vendoorDatabaseSpecification() = database {
-    version { "1.0.0" }
-
-    collections {
-        collection {
-            name { "User" }
-        }
-    }
-}
